backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
supabase>=2.10.0
pydantic>=2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
PyJWT>=2.8.0
psycopg2-binary==2.9.9
django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
email-validator>=2.0.0
websockets==12.0
